Remove indirection in `possible_moves`
It used the filter method on an iterator but this required a function
item to be cast to a function pointer which loses information for the
compiler. Now the filtering is node in the next method where the filter
function doesn't need to be stored so it doesn't need to be cast to a
function pointer.